azure_function/requirements.txt
azure-functions
azure-kusto-data==4.*
azure-kusto-ingest==4.*
azure-identity
requests
pandas
